Separate utils to different files under new src/ directory (#13)
