+++ src/paths.py
@@ -1,0 +1,29 @@
+from pathlib import Path
+
+PROJECT_DIRECTORY = Path(__file__).parent.parent
+
+DATA_PATH = PROJECT_DIRECTORY /  'data'
+DATA_SPLIT_PATH = DATA_PATH / 'split'
+DATA_RAW_PATH = DATA_PATH / 'raw'
+
+KEY_GENRE_ENGLISH_PATH = DATA_PATH / 'key_genre_english.csv'
+KEY_ENGLISH_PATH = DATA_PATH / 'key_english.csv'
+TABLE_KEY_PATH = DATA_PATH / 't_key.csv'
+
+TABLE_DIRECTORY = DATA_PATH
+
+TABLE_NAME_FORMAT = '{table}.csv'
+SPLIT_DATASET_FORMAT = '{table}_{dataset}.txt'
+
+HELSINKI_RAW_PATH = DATA_RAW_PATH / 'helsinki'
+HELSINKI_RAW_TAR_PATH = f'{HELSINKI_RAW_PATH}.tar.gz'
+MIDDLE_ENGLISH_PROSE_VERSE_RAW_PATH = DATA_RAW_PATH / 'middle_english_prose'
+MIDDLE_ENGLISH_PROSE_VERSE_RAW_TAR_PATH = f'{MIDDLE_ENGLISH_PROSE_VERSE_RAW_PATH}.tar.gz'
+WEST_SAXON_GOSPEL_CSV_PATH = DATA_PATH / 't_ws.csv'
+
+WYCLIFFE_DIRECTORY_PATH = DATA_PATH / 'wycbible'
+WYCLIFFE_KEY_PATH = WYCLIFFE_DIRECTORY_PATH / 'index.txt'
+WYCLIFFE_CSV_PATH = DATA_PATH / 't_wyc.csv'
+
+AELFRIC_OLD_TESTAMENT_XML_PATH = HELSINKI_RAW_PATH / "0_AelfricTheOldTestament.xml"
+AELFRIC_CSV_PATH = DATA_PATH / "t_alf.csv"
